Moved flatten and related functions to Importer class
