+++ src/api/libcellml/importer.h
@@ -61,11 +61,28 @@ public:
      */
     static void resolveImports(const std::string &baseFile, const ModelPtr &model);
 
+    /**
+     * @brief Flatten this model.
+     *
+     * Instantiates all imports and removes them from this model.
+     * The result is a self-contained model requiring no external
+     * resources and having no imports.
+     *
+     * The effects of this method cannot be undone.
+     *
+     * @sa clone
+     */
+    void flatten(ModelPtr &model);
+
 private:
     Importer(); /**< Constructor */
 
     struct ImporterImpl; /**< Forward declaration for pImpl idiom. */
     ImporterImpl *mPimpl; /**< Private member to implementation pointer. */
 };
+
+// void recordVariableEquivalences(const ComponentPtr &component, EquivalenceMap &equivalenceMap, IndexStack &indexStack);
+// void generateEquivalenceMap(const ComponentPtr &component, EquivalenceMap &map, IndexStack &indexStack);
+// void applyEquivalenceMapToModel(const EquivalenceMap &map, const ModelPtr &model);
 
 } // namespace libcellml
